Added the pantry and shoppinglist tables in models and moved over spoonacular.js to public.
