+++ controllers/pantreeController.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const router = express.Router();
 
-const spoonacular = require("../models/spoonacular.js");
+const spoonacular = require("./../public/assets/js/spoonacular");
 
 // Create all our routes and set up logic within those routes where required.
 router.get("/", (req, res) => {
