+++ public/assets/js/spoonacular.js
